Fix CodeFactor issues
